+++ sdk/data_cosmos/src/operations/get_document.rs
@@ -52,6 +52,10 @@ impl GetDocumentBuilder {
             if let Some(cl) = &self.consistency_level {
                 request.insert_headers(cl);
             }
+            crate::cosmos_entity::add_as_partition_key_header_serialized2(
+                self.client.partition_key_serialized(),
+                &mut request,
+            );
 
             request.set_body(azure_core::EMPTY_BODY.into());
