+++ test/transactions/document_operations/0_request.json
@@ -1,0 +1,1 @@
+{"uri":"/dbs","method":"POST","headers":{"authorization":"<<STRIPPED>>","user-agent":"azsdk-rust-data_cosmos/0.1.0 (1.59.0; linux; x86_64)","x-ms-date":"Tue, 29 Mar 2022 13:57:03 GMT","x-ms-version":"2018-12-31"},"body":"eyJpZCI6InRlc3QtY29zbW9zLWRiLWNyZWF0ZS1hbmQtZGVsZXRlLWRvY3VtZW50In0="}
