+++ test/transactions/document_operations/0_response.json
@@ -1,0 +1,1 @@
+{"status":201,"headers":{"cache-control":"no-store, no-cache","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:00 GMT","etag":"\"00003101-0000-0100-0000-6243102d0000\"","lsn":"54","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","transfer-encoding":"chunked","x-ms-activity-id":"fe5542d4-5798-4d1c-93d2-5c9ac232e083","x-ms-cosmos-llsn":"54","x-ms-cosmos-quorum-acked-llsn":"53","x-ms-current-replica-set-size":"4","x-ms-current-write-quorum":"3","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"53","x-ms-last-state-change-utc":"Wed, 23 Mar 2022 20:04:36.240 GMT","x-ms-number-of-read-regions":"0","x-ms-quorum-acked-lsn":"53","x-ms-request-charge":"4.95","x-ms-request-duration-ms":"23.579","x-ms-resource-quota":"databases=1000;","x-ms-resource-usage":"databases=2;","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#54","x-ms-transport-request-id":"96734","x-ms-xp-role":"1"},"body":"eyJpZCI6InRlc3QtY29zbW9zLWRiLWNyZWF0ZS1hbmQtZGVsZXRlLWRvY3VtZW50IiwiX3JpZCI6IllPdDFBQT09IiwiX3NlbGYiOiJkYnNcL1lPdDFBQT09XC8iLCJfZXRhZyI6IlwiMDAwMDMxMDEtMDAwMC0wMTAwLTAwMDAtNjI0MzEwMmQwMDAwXCIiLCJfY29sbHMiOiJjb2xsc1wvIiwiX3VzZXJzIjoidXNlcnNcLyIsIl90cyI6MTY0ODU2MjIyMX0="}
