+++ test/transactions/document_operations/10_response.json
@@ -1,0 +1,1 @@
+{"status":204,"headers":{"cache-control":"no-store, no-cache","content-length":"0","content-location":"https://azuresdktesting2.documents.azure.com/dbs/test-cosmos-db-create-and-delete-document","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:05 GMT","lsn":"56","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","x-ms-activity-id":"094e0941-0510-46d0-a21d-cf1280ae1074","x-ms-cosmos-llsn":"56","x-ms-cosmos-quorum-acked-llsn":"55","x-ms-current-replica-set-size":"4","x-ms-current-write-quorum":"3","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"55","x-ms-last-state-change-utc":"Wed, 23 Mar 2022 20:04:36.240 GMT","x-ms-number-of-read-regions":"0","x-ms-quorum-acked-lsn":"55","x-ms-request-charge":"4.95","x-ms-request-duration-ms":"27.817","x-ms-resource-quota":"databases=1000;","x-ms-resource-usage":"databases=1;","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#56","x-ms-transport-request-id":"96739","x-ms-xp-role":"1"},"body":""}
