+++ test/transactions/document_operations/1_request.json
@@ -1,0 +1,1 @@
+{"uri":"/dbs/test-cosmos-db-create-and-delete-document/colls","method":"POST","headers":{"authorization":"<<STRIPPED>>","user-agent":"azsdk-rust-data_cosmos/0.1.0 (1.59.0; linux; x86_64)","x-ms-date":"Tue, 29 Mar 2022 13:57:03 GMT","x-ms-offer-throughput":"400","x-ms-version":"2018-12-31"},"body":"eyJpZCI6InRlc3QtY29sbGVjdGlvbi1jcmVhdGUtYW5kLWRlbGV0ZS1kb2N1bWVudCIsImluZGV4aW5nUG9saWN5Ijp7ImF1dG9tYXRpYyI6dHJ1ZSwiaW5kZXhpbmdNb2RlIjoiY29uc2lzdGVudCIsImluY2x1ZGVkUGF0aHMiOltdLCJleGNsdWRlZFBhdGhzIjpbXX0sInBhcnRpdGlvbktleSI6eyJwYXRocyI6WyIvaWQiXSwia2luZCI6Ikhhc2gifX0="}
