+++ test/transactions/document_operations/1_response.json
@@ -1,0 +1,1 @@
+{"status":201,"headers":{"cache-control":"no-store, no-cache","collection-partition-index":"0","collection-service-index":"0","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:03 GMT","etag":"\"00003301-0000-0100-0000-624310300000\"","lsn":"1","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","transfer-encoding":"chunked","x-ms-activity-id":"aac72148-2135-4d46-8c29-abe2fb3bd581","x-ms-alt-content-path":"dbs/test-cosmos-db-create-and-delete-document","x-ms-cosmos-item-llsn":"1","x-ms-cosmos-llsn":"1","x-ms-cosmos-quorum-acked-llsn":"1","x-ms-current-replica-set-size":"4","x-ms-current-write-quorum":"3","x-ms-documentdb-partitionkeyrangeid":"0","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"1","x-ms-item-lsn":"1","x-ms-last-state-change-utc":"Tue, 29 Mar 2022 13:45:10.219 GMT","x-ms-number-of-read-regions":"0","x-ms-quorum-acked-lsn":"1","x-ms-request-charge":"1","x-ms-request-duration-ms":"0.445","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#1","x-ms-transport-request-id":"2","x-ms-xp-role":"1"},"body":"eyJpZCI6InRlc3QtY29sbGVjdGlvbi1jcmVhdGUtYW5kLWRlbGV0ZS1kb2N1bWVudCIsImluZGV4aW5nUG9saWN5Ijp7ImluZGV4aW5nTW9kZSI6ImNvbnNpc3RlbnQiLCJhdXRvbWF0aWMiOnRydWUsImluY2x1ZGVkUGF0aHMiOlt7InBhdGgiOiJcLyoifV0sImV4Y2x1ZGVkUGF0aHMiOlt7InBhdGgiOiJcL1wiX2V0YWdcIlwvPyJ9XX0sInBhcnRpdGlvbktleSI6eyJwYXRocyI6WyJcL2lkIl0sImtpbmQiOiJIYXNoIn0sImNvbmZsaWN0UmVzb2x1dGlvblBvbGljeSI6eyJtb2RlIjoiTGFzdFdyaXRlcldpbnMiLCJjb25mbGljdFJlc29sdXRpb25QYXRoIjoiXC9fdHMiLCJjb25mbGljdFJlc29sdXRpb25Qcm9jZWR1cmUiOiIifSwiZ2Vvc3BhdGlhbENvbmZpZyI6eyJ0eXBlIjoiR2VvZ3JhcGh5In0sIl9yaWQiOiJZT3QxQVBLWTlrUT0iLCJfdHMiOjE2NDg1NjIyMjQsIl9zZWxmIjoiZGJzXC9ZT3QxQUE9PVwvY29sbHNcL1lPdDFBUEtZOWtRPVwvIiwiX2V0YWciOiJcIjAwMDAzMzAxLTAwMDAtMDEwMC0wMDAwLTYyNDMxMDMwMDAwMFwiIiwiX2RvY3MiOiJkb2NzXC8iLCJfc3Byb2NzIjoic3Byb2NzXC8iLCJfdHJpZ2dlcnMiOiJ0cmlnZ2Vyc1wvIiwiX3VkZnMiOiJ1ZGZzXC8iLCJfY29uZmxpY3RzIjoiY29uZmxpY3RzXC8ifQ=="}
