+++ test/transactions/document_operations/2_request.json
@@ -1,0 +1,1 @@
+{"uri":"/dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document/docs","method":"POST","headers":{"authorization":"<<STRIPPED>>","user-agent":"azsdk-rust-data_cosmos/0.1.0 (1.59.0; linux; x86_64)","x-ms-cosmos-allow-tentative-writes":"false","x-ms-date":"Tue, 29 Mar 2022 13:57:06 GMT","x-ms-documentdb-is-upsert":"false","x-ms-documentdb-partitionkey":"[\"test-document-name-create-and-delete-document\"]","x-ms-version":"2018-12-31"},"body":"eyJpZCI6InRlc3QtZG9jdW1lbnQtbmFtZS1jcmVhdGUtYW5kLWRlbGV0ZS1kb2N1bWVudCIsImhlbGxvIjo0Mn0="}
