+++ test/transactions/document_operations/3_request.json
@@ -1,0 +1,1 @@
+{"uri":"/dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document/docs","method":"GET","headers":{"authorization":"<<STRIPPED>>","user-agent":"azsdk-rust-data_cosmos/0.1.0 (1.59.0; linux; x86_64)","x-ms-date":"Tue, 29 Mar 2022 13:57:07 GMT","x-ms-max-item-count":"-1","x-ms-version":"2018-12-31"},"body":""}
