+++ test/transactions/document_operations/4_response.json
@@ -1,0 +1,1 @@
+{"status":200,"headers":{"cache-control":"no-store, no-cache","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:04 GMT","lsn":"2","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","transfer-encoding":"chunked","x-ms-activity-id":"cbaedb3d-5e5c-46a6-b0fa-1cee29ea946e","x-ms-alt-content-path":"dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document","x-ms-content-path":"YOt1APKY9kQ=","x-ms-cosmos-is-partition-key-delete-pending":"false","x-ms-cosmos-llsn":"2","x-ms-cosmos-query-execution-info":"{\"reverseRidEnabled\":false,\"reverseIndexScan\":false}","x-ms-documentdb-partitionkeyrangeid":"0","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"2","x-ms-item-count":"1","x-ms-last-state-change-utc":"Tue, 29 Mar 2022 13:45:18.584 GMT","x-ms-number-of-read-regions":"0","x-ms-request-charge":"2.26","x-ms-request-duration-ms":"0.428","x-ms-resource-quota":"documentSize=10240;documentsSize=10485760;documentsCount=-1;collectionSize=10485760;","x-ms-resource-usage":"documentSize=0;documentsSize=1;documentsCount=1;collectionSize=1;","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#2","x-ms-transport-request-id":"1","x-ms-xp-role":"2"},"body":"eyJfcmlkIjoiWU90MUFQS1k5a1E9IiwiRG9jdW1lbnRzIjpbeyJpZCI6InRlc3QtZG9jdW1lbnQtbmFtZS1jcmVhdGUtYW5kLWRlbGV0ZS1kb2N1bWVudCIsImhlbGxvIjo0MiwiX3JpZCI6IllPdDFBUEtZOWtRQkFBQUFBQUFBQUE9PSIsIl9zZWxmIjoiZGJzXC9ZT3QxQUE9PVwvY29sbHNcL1lPdDFBUEtZOWtRPVwvZG9jc1wvWU90MUFQS1k5a1FCQUFBQUFBQUFBQT09XC8iLCJfZXRhZyI6IlwiMWIwMDM0N2ItMDAwMC0wMTAwLTAwMDAtNjI0MzEwMzAwMDAwXCIiLCJfYXR0YWNobWVudHMiOiJhdHRhY2htZW50c1wvIiwiX3RzIjoxNjQ4NTYyMjI0fV0sIl9jb3VudCI6MX0="}
