+++ test/transactions/document_operations/5_response.json
@@ -1,0 +1,1 @@
+{"status":200,"headers":{"cache-control":"no-store, no-cache","content-location":"https://azuresdktesting2.documents.azure.com/dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document/docs/test-document-name-create-and-delete-document","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:04 GMT","etag":"\"1b00347b-0000-0100-0000-624310300000\"","lsn":"2","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","transfer-encoding":"chunked","x-ms-activity-id":"632556b5-db99-40a6-a496-21de54657b12","x-ms-alt-content-path":"dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document","x-ms-content-path":"YOt1APKY9kQ=","x-ms-cosmos-item-llsn":"2","x-ms-cosmos-llsn":"2","x-ms-cosmos-quorum-acked-llsn":"2","x-ms-current-replica-set-size":"4","x-ms-current-write-quorum":"3","x-ms-documentdb-partitionkeyrangeid":"0","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"2","x-ms-item-lsn":"2","x-ms-last-state-change-utc":"Tue, 29 Mar 2022 13:45:10.219 GMT","x-ms-number-of-read-regions":"0","x-ms-quorum-acked-lsn":"2","x-ms-request-charge":"1","x-ms-request-duration-ms":"0.254","x-ms-resource-quota":"documentSize=10240;documentsSize=10485760;documentsCount=-1;collectionSize=10485760;","x-ms-resource-usage":"documentSize=0;documentsSize=1;documentsCount=1;collectionSize=1;","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#2","x-ms-transport-request-id":"5","x-ms-xp-role":"1"},"body":"eyJpZCI6InRlc3QtZG9jdW1lbnQtbmFtZS1jcmVhdGUtYW5kLWRlbGV0ZS1kb2N1bWVudCIsImhlbGxvIjo0MiwiX3JpZCI6IllPdDFBUEtZOWtRQkFBQUFBQUFBQUE9PSIsIl9zZWxmIjoiZGJzXC9ZT3QxQUE9PVwvY29sbHNcL1lPdDFBUEtZOWtRPVwvZG9jc1wvWU90MUFQS1k5a1FCQUFBQUFBQUFBQT09XC8iLCJfZXRhZyI6IlwiMWIwMDM0N2ItMDAwMC0wMTAwLTAwMDAtNjI0MzEwMzAwMDAwXCIiLCJfYXR0YWNobWVudHMiOiJhdHRhY2htZW50c1wvIiwiX3RzIjoxNjQ4NTYyMjI0fQ=="}
