+++ test/transactions/document_operations/7_response.json
@@ -1,0 +1,1 @@
+{"status":200,"headers":{"cache-control":"no-store, no-cache","content-location":"https://azuresdktesting2.documents.azure.com/dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document/docs/test-document-name-create-and-delete-document","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:05 GMT","etag":"\"1b00357b-0000-0100-0000-624310310000\"","lsn":"3","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","transfer-encoding":"chunked","x-ms-activity-id":"8046ed4f-c810-4ffb-8dc1-bf3cdb507a69","x-ms-alt-content-path":"dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document","x-ms-content-path":"YOt1APKY9kQ=","x-ms-cosmos-item-llsn":"3","x-ms-cosmos-llsn":"3","x-ms-documentdb-partitionkeyrangeid":"0","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"3","x-ms-item-lsn":"3","x-ms-last-state-change-utc":"Tue, 29 Mar 2022 13:45:18.585 GMT","x-ms-number-of-read-regions":"0","x-ms-request-charge":"1","x-ms-request-duration-ms":"0.248","x-ms-resource-quota":"documentSize=10240;documentsSize=10485760;documentsCount=-1;collectionSize=10485760;","x-ms-resource-usage":"documentSize=0;documentsSize=1;documentsCount=1;collectionSize=1;","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#3","x-ms-transport-request-id":"1","x-ms-xp-role":"1"},"body":"eyJpZCI6InRlc3QtZG9jdW1lbnQtbmFtZS1jcmVhdGUtYW5kLWRlbGV0ZS1kb2N1bWVudCIsImhlbGxvIjoxOTAsIl9yaWQiOiJZT3QxQVBLWTlrUUJBQUFBQUFBQUFBPT0iLCJfc2VsZiI6ImRic1wvWU90MUFBPT1cL2NvbGxzXC9ZT3QxQVBLWTlrUT1cL2RvY3NcL1lPdDFBUEtZOWtRQkFBQUFBQUFBQUE9PVwvIiwiX2V0YWciOiJcIjFiMDAzNTdiLTAwMDAtMDEwMC0wMDAwLTYyNDMxMDMxMDAwMFwiIiwiX2F0dGFjaG1lbnRzIjoiYXR0YWNobWVudHNcLyIsIl90cyI6MTY0ODU2MjIyNX0="}
