+++ test/transactions/document_operations/8_response.json
@@ -1,0 +1,1 @@
+{"status":204,"headers":{"cache-control":"no-store, no-cache","content-length":"0","content-location":"https://azuresdktesting2.documents.azure.com/dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document/docs/test-document-name-create-and-delete-document","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:05 GMT","lsn":"4","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","x-ms-activity-id":"f0f00883-de96-466f-b9df-a8a71767e649","x-ms-alt-content-path":"dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document","x-ms-content-path":"YOt1APKY9kQ=","x-ms-cosmos-llsn":"4","x-ms-cosmos-quorum-acked-llsn":"3","x-ms-current-replica-set-size":"4","x-ms-current-write-quorum":"3","x-ms-documentdb-partitionkeyrangeid":"0","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"3","x-ms-last-state-change-utc":"Tue, 29 Mar 2022 13:45:10.219 GMT","x-ms-number-of-read-regions":"0","x-ms-quorum-acked-lsn":"3","x-ms-request-charge":"5.9","x-ms-request-duration-ms":"4.544","x-ms-resource-quota":"documentSize=10240;documentsSize=10485760;documentsCount=-1;collectionSize=10485760;","x-ms-resource-usage":"documentSize=0;documentsSize=0;documentsCount=1;collectionSize=0;","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#4","x-ms-transport-request-id":"7","x-ms-xp-role":"1"},"body":""}
