+++ test/transactions/document_operations/9_response.json
@@ -1,0 +1,1 @@
+{"status":200,"headers":{"cache-control":"no-store, no-cache","content-location":"https://azuresdktesting2.documents.azure.com/dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document/docs","content-type":"application/json","date":"Tue, 29 Mar 2022 13:57:05 GMT","lsn":"4","pragma":"no-cache","server":"Microsoft-HTTPAPI/2.0","strict-transport-security":"max-age=31536000","transfer-encoding":"chunked","x-ms-activity-id":"9a1ff7e5-603b-484a-b462-8170c9e9d7cd","x-ms-alt-content-path":"dbs/test-cosmos-db-create-and-delete-document/colls/test-collection-create-and-delete-document","x-ms-content-path":"YOt1APKY9kQ=","x-ms-cosmos-llsn":"4","x-ms-documentdb-partitionkeyrangeid":"0","x-ms-gatewayversion":"version=2.14.0","x-ms-global-committed-lsn":"4","x-ms-item-count":"0","x-ms-last-state-change-utc":"Tue, 29 Mar 2022 13:45:20.882 GMT","x-ms-number-of-read-regions":"0","x-ms-request-charge":"1","x-ms-request-duration-ms":"0.488","x-ms-resource-quota":"documentSize=10240;documentsSize=10485760;documentsCount=-1;collectionSize=10485760;","x-ms-resource-usage":"documentSize=0;documentsSize=0;documentsCount=0;collectionSize=0;","x-ms-schemaversion":"1.13","x-ms-serviceversion":"version=2.14.0.0","x-ms-session-token":"0:-1#4","x-ms-transport-request-id":"1","x-ms-xp-role":"1"},"body":"eyJfcmlkIjoiWU90MUFQS1k5a1E9IiwiRG9jdW1lbnRzIjpbXSwiX2NvdW50IjowfQ=="}
